test: share simulation helper
