+++ tests/simulation/exclusive-gateway.multiple.test.js
@@ -1,38 +1,6 @@
 const { test } = require('node:test');
 const assert = require('assert');
-const fs = require('fs');
-const path = require('path');
-const vm = require('vm');
-
-function loadSimulation() {
-  const sandbox = {
-    console,
-    setTimeout,
-    clearTimeout,
-    localStorage: {
-      _data: {},
-      getItem(key) { return this._data[key] || null; },
-      setItem(key, val) { this._data[key] = String(val); },
-      removeItem(key) { delete this._data[key]; }
-    }
-  };
-  const streamCode = fs.readFileSync(path.resolve(__dirname, '../../public/js/core/stream.js'), 'utf8');
-  const simulationCode = fs.readFileSync(path.resolve(__dirname, '../../public/js/core/simulation.js'), 'utf8');
-  vm.runInNewContext(streamCode, sandbox);
-  vm.runInNewContext(simulationCode, sandbox);
-  return sandbox.createSimulation;
-}
-
-function createSimulationInstance(elements, opts = {}) {
-  const map = new Map(elements.map(e => [e.id, e]));
-  const elementRegistry = {
-    get(id) { return map.get(id); },
-    filter(fn) { return Array.from(map.values()).filter(fn); }
-  };
-  const canvas = { addMarker() {}, removeMarker() {} };
-  const createSimulation = loadSimulation();
-  return createSimulation({ elementRegistry, canvas }, opts);
-}
+const { createSimulationInstance } = require('../helpers/simulation');
 
 function buildMultipleConditionalDiagram() {
   const start = { id: 'start', type: 'bpmn:StartEvent', outgoing: [], incoming: [], businessObject: { $type: 'bpmn:StartEvent' } };
